+++ NEWS.md
@@ -1,5 +1,6 @@
 # csm (development version)
 * Added basic functions for data quality check
+* New function to select a center from the output given by `nest_tables`
 * Fix `.travis.yml`
 
 # csm 0.0.0.9000